feat: Score 수정

+++ Assets/08_Scripts/UI/Score/IScorable.cs
@@ -16,7 +16,7 @@ public enum ScoreType
 
 public interface IScorable
 {
-    void SetScore(ScoreType scoreType, float score);
+    void SetScore(ScoreType scoreType, int score);
 
     float GetScore(ScoreType scoreType);
