+++ Assets/08_Scripts/UI/Score/ScoreManager.cs
@@ -4,14 +4,14 @@ using UnityEngine;
 
 public class ScoreManager : MonoBehaviour, IScorable
 {
-    Dictionary<ScoreType, float> dicScore = new Dictionary<ScoreType, float>();
+    Dictionary<ScoreType, int> dicScore = new Dictionary<ScoreType, int>();
 
     public float GetScore(ScoreType scoreType)
     {
         return dicScore[scoreType];
     }
 
-    public void SetScore(ScoreType scoreType, float score)
+    public void SetScore(ScoreType scoreType, int score)
     {
         dicScore.Add(scoreType, score);
     }
